translation

+++ app/Http/Controllers/Api/ItemsController.php
@@ -4,21 +4,17 @@ namespace App\Http\Controllers\Api;
 
 use App\Item;
 use Illuminate\Http\Request;
-use App\Exceptions\Api\ApiException;
+use Spatie\QueryBuilder\Filter;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\ItemResource;
+use Spatie\QueryBuilder\QueryBuilder;
 use Illuminate\Support\Facades\Validator;
 
 class ItemsController extends Controller
 {
     public function index()
     {
-        return ItemResource::collection(auth('api')->user()->items);
-    }
-
-    public function show(Item $item)
-    {
-        return new ItemResource($item);
+        return ItemResource::collection(Item::all());
     }
 
     public function store(Request $request)
@@ -30,21 +26,38 @@ class ItemsController extends Controller
             'color_id' => ['required', 'exists:colors,id'],
             'model_id' => ['required', 'exists:models,id'],
             'brand_id' => ['required', 'exists:brands,id'],
+
         ]);
 
         if ($validate_request->fails()) {
-            throw new ApiException($validate_request->errors()->first(), 400);
+            $this->addMultibleResponse($validate_request->errors())->addStatusCode(400);
+            return $this->response();
         }
-
         return (new Item)->createItem($request);
+    }
+
+    public function show(Item $item)
+    {
+        return new ItemResource($item);
+    }
+
+    public function update(Request $request, $id)
+    {
+        //
     }
 
     public function destroy(Item $item)
     {
-        $item->delete();
-
-        $this->addResponse(trans('posts.successfully_deleted'))->addStatusCode(200);
-
-        return  $this->response();
+        $user = auth('api')->user();
+        if ($user->can('destroy', $item)) {
+            $item->delete();
+            $this->addResponse(trans('messages.successfully_deleted'))->addStatusCode(200);
+            return  $this->response();
+        }
+        throw new ApiException(trans('auth.not_authorized'), 400);
+    }
+    public function userItems()
+    {
+        return  ItemResource::collection(auth('api')->user()->items()->get());
     }
 }
